Move the PERFT test to backend to avoid building unit-tests into the application.
