+++ Terminal/Terminal.csproj
@@ -15,7 +15,6 @@
 
     <ItemGroup>
       <ProjectReference Include="..\Backend\Backend.csproj" />
-      <ProjectReference Include="..\Test\Test.csproj" />
     </ItemGroup>
 
     <ItemGroup>
